Adapt to the model

+++ dataset.py
@@ -38,7 +38,7 @@ class PawpularityDataset(Dataset):
         # swap color axis because
         # numpy image: H x W x C
         # torch image: C x H x W
-        image = image.transpose((2, 0, 1))
+        image = image.transpose((2, 0, 1)).astype(np.float32)
         sample = {'image': torch.from_numpy(image), 'score': torch.LongTensor(score)}
 
         if self.transform:
